redamodif

+++ src/main/java/com/codingtech/formationcenter/repo/PromotionRepo.java
@@ -3,5 +3,5 @@ package com.codingtech.formationcenter.repo;
 import com.codingtech.formationcenter.entity.Promotion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromotionRepo  extends JpaRepository<Promotion,Integer> {
+public interface PromotionRepo  extends JpaRepository<Promotion,Long> {
 }
